requirements.txt
pycryptodome
lxml
prettytable
